fix foreign keys: when using Long our migrations need to use BIGINT

+++ src/main/resources/db/migration/V2__Create_categories_table.sql
@@ -1,7 +1,8 @@
 -- V2__Create_categories_table
 
 CREATE TABLE categories (
-    id INT AUTO_INCREMENT PRIMARY KEY,
+    id BIGINT AUTO_INCREMENT PRIMARY KEY,
     description VARCHAR(100) NOT NULL,
     created_at TIMESTAMP DEFAULT CURRENT_TIMESTAMP NOT NULL
 );
+
